⛺️📉 Checkpoint ./komutlar/oylama.js:22384402/3153

+++ komutlar/oylama.js
@@ -1,48 +1,42 @@
-const Discord = require('discord.js');
+const Discord = require("discord.js");
 
 exports.run = (client, message, args) => {
-message.delete();
+  if (message.channel.type == "dm") return;
+  if (message.channel.type !== "text") return;
 
-let question = args.join(' ');
+  if (!message.member.hasPermission("MANAGE_MESSAGES")) return message.reply(`Bu Komutu Kullanabilmek İçin **Mesajları Yönet** İznine Sahip Olmalısın Ama Senin Yok Kullanamazsın`).then(m => m.delete({ timeout: 10000}));
 
-let user = message.author.username
+  message.delete();
 
-if (!question) return message.channel.sendEmbed(
+  let köseoylama = args.join(" ");
 
-new Discord.RichEmbed()
+  let user = message.author.username;
 
-.addField(`:x: | Yazı yazman gerek.`)).then(m => m.delete(5000));
+  if (!köseoylama) return message.channel.send(new Discord.MessageEmbed().setTitle(`Yazı Yazmayı Unuttun Kanka`)).then(m => m.delete(({ timeout: 5000})));
 
-console.log("Oylama Komutu " + message.author.username + '#' + message.author.discriminator + " Tarafından Kullanıldı.")
-message.channel.sendEmbed(
-
-new Discord.RichEmbed()
-
-.setColor("RED")
-.setThumbnail(client.user.avatarURL)
-.setTimestamp()
-.setFooter('Oyun Elitleri Oylama Sistemi', client.user.avatarURL)
-
-.addField(`**Oylama**`, `**${question}**`)).then(function(message) {
-
-message.react('✅');
-
-message.react('❌');
-
-});
-
+  message.channel.send(new Discord.MessageEmbed()
+        .setColor("PURPLE")
+        .addField(`Oylama Yapınız`, `**${köseoylama}**`)
+        .addField(`Komutu Kullanan Kişi ${message.author.tag}`,`Evet için <:yr_evet:793837194175447090> Hayır için <:yr_hayr:793837203478020127> Oylayınız.`)
+        .setDescription(` Not : Lütfen 2 Emojiyide Basmayınız`)
+        .setThumbnail(client.user.avatarURL())
+        .setFooter("Yrnex Oylama Sistemi", client.user.avatarURL())
+    )
+    .then(function(message) {
+    message.react(`<:yr_evet:793837194175447090>`);
+    message.react(`:yr_hayr:`);
+    });
 };
-//Oyun Elitleri Botuna Aittir İzinsiz Paylaşılması Yasaktır!!!
-exports.conf = {
-enabled: true,
-guildOnly: false,
-aliases: ['oylama'],
 
-permLevel: 2
+exports.conf = {
+  enabled: true,
+  guildOnly: false,
+  aliases: ["oylama"],
+  permLevel: 0
 };
 
 exports.help = {
-name: 'oylama',
-description: 'Oylama yapmanızı sağlar.',
-usage: 'oylama <oylamaismi>'
-}
+  name: "oylama",
+  description: "yardım menüsü.",
+  usage: "yardım"
+};
